refactor(Invalidate): deprecate icon
