+++ src/ui/svg/icons/Invalidate_deprecated.tsx
